refreshing alerts during refresh of favorites and NTA results
